requirements.txt

---

**requirements.txt**

```plaintext
numpy
pytesseract
Pillow
torch
datasets
transformers
protobuf
tensorflow
sentencepiece
opencv-python
rapidfuzz
lxml
